requirements.txt
torch
argparse
Pillow
torchvision
mock
albumentations
pandas
numpy
opencv-python
matplotlib
Cython
pycocotools
cv2
tqdm
tensorboard=2.1.0